+++ apps/hub/src/app/portals/eutthan-admin/[[...slug]]/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { EutthanPortal } from "@/components/eutthan/eutthan-portal";
+
+export default function EutthanPage() {
+  return <EutthanPortal />;
+}